+++ src/client/java/doctor_m/DOCTORMClient.java
@@ -1,11 +1,13 @@
 package doctor_m;
 
+import client.render.TitanDimensionEffects;
 import client.render.entity.evereye_renderer;
 import client.render.entity.tardis_renderer;
 import doctor_m.entities.entities;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
+import net.fabricmc.fabric.api.client.rendering.v1.DimensionRenderingRegistry;
 import net.fabricmc.fabric.api.client.rendering.v1.EntityRendererRegistry;
 import net.minecraft.client.render.entity.model.EntityModelLayer;
 import net.minecraft.util.Identifier;
@@ -18,6 +20,12 @@ public class DOCTORMClient implements ClientModInitializer {
 
     @Override
     public void onInitializeClient() {
+        // 泰坦维度效果
+        DimensionRenderingRegistry.registerDimensionEffects(
+                new Identifier("doctor_m", "titan"),
+                new TitanDimensionEffects()
+        );
+
         EntityRendererRegistry.register(entities.TYPE_103_TARDIS, tardis_renderer::new);
         EntityRendererRegistry.register(entities.TYPE_103W_EVEREYE, evereye_renderer::new);
     }