+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## [ v2.1 ]
 
+### 新维度 / New Dimension
+
+- 木卫六-泰坦 / Saturn VI-Titan
+
 ### 新翻译 / New lang
 
 -  空